Allow for template extension + blocks
What?
=====

It's common to have base templates where other fragments can inject
content into them. This might be a template for blog posts, interior
pages, or other types of content that is reused but needs to allow for
subbing in not just values, but additional markup.

This introduces fragment "subclassing" through two new constructs:
extend and block.

// fragment/page.skt
.page
  %header
    %h2= block header
    - block content

// fragment/readme.skt
- extend fragment/page.skt
- block header
  README
- block content
  %h3 Welcome!

  %p If you're reading this for the first time: hello there!