Extract string management into a Builder to aggregate warnings
What?
=====

Throughout the process of traversing the parsed nodes AST, operations
can be problematic for various reasons. A for loop may attempt to
iterate over a value from JSON that's not present (or not an Array, as
we don't currently allow for iteration over objects). It might fail
because it's attempting to access a misnamed fragment, or an invalid
block name.

Rather than continue on without knowing (having certain sections parse
as an empty string), this introduces a simple structure to aggregate the
list of warnings for future use.