Test fragments and improve functionality around finding fragments
What?
=====

This introduces tests for fragments and improves behavior, making it
more straightforward to reference fragments by removing the necessary
prefix.

Previously, the behavior only allowed for fragments within the path
`./fragments`, and required referencing fragments with the full path:

```
- fragment fragments/thing.skt
```

This now allows for an optional flag to be provided (still defaulting to
`./fragments`) and allows referencing paths within the calculated
fragment directory, e.g.

```
- fragment thing.skt
```